Mark second milestone as complete.
I haven't pushed all the possible conversions from the cdli
database dump because it's a lot of data and I expect to make
a several more markup changes.

Nevertheless I added the BM--PJ tablet set which is 589 documents
to the original 4 test conversions, and added instructions for how
to update the collection to the cdli-cts README. This is a fair
start, I think, without wasting a lot of space in the repository
history.